share db connection
